+++ manta-accounting/src/key.rs
@@ -31,7 +31,7 @@ use core::{
 };
 use manta_crypto::{
     key::KeyDerivationFunction,
-    rand::{CryptoRng, RngCore, Sample},
+    rand::{RngCore, Sample},
 };
 use manta_util::collections::btree_map::{self, BTreeMap};
 
@@ -269,7 +269,7 @@ where
     #[inline]
     fn sample<R>(distribution: (D, F), rng: &mut R) -> Self
     where
-        R: CryptoRng + RngCore + ?Sized,
+        R: RngCore + ?Sized,
     {
         Self::new(H::sample(distribution.0, rng), distribution.1)
     }